Fix [Feature set] The URL field for external target should have predefined options
