+++ src/components/FeatureSetsPanel/FeatureSetsPanel.js
@@ -58,6 +58,7 @@ const FeatureSetsPanel = ({
     isAccessKeyValid: true
   })
   const [disableButtons, setDisableButtons] = useState({
+    isExternalOfflineTargetPathEditModeClosed: true,
     isOnlineTargetPathEditModeClosed: true,
     isOfflineTargetPathEditModeClosed: true,
     isUrlEditModeClosed: true
