+++ src/components/FeatureSetsPanel/FeatureSetsPanelTargetStore/FeatureSetsPanelTargetStore.js
@@ -42,6 +42,8 @@ import {
   ONLINE
 } from './featureSetsPanelTargetStore.util'
 
+import { isUrlInputValid } from '../UrlPath.utils'
+
 const FeatureSetsPanelTargetStore = ({
   featureStore,
   project,
@@ -254,22 +256,64 @@ const FeatureSetsPanelTargetStore = ({
     )
   }
 
-  const handleExternalOfflineKindPathOnBlur = event => {
-    const target = featureStore.newFeatureSet.spec.targets.find(
-      targetKind => targetKind.name === EXTERNAL_OFFLINE
-    )
+  const handleUrlSelectOnChange = () => {
+    setValidation(state => ({
+      ...state,
+      isExternalOfflineTargetPathValid: true
+    }))
 
-    if (event.target.value !== target.path) {
+    setNewFeatureSetTarget(
+      featureStore.newFeatureSet.spec.targets.map(targetKind => {
+        if (targetKind.name === EXTERNAL_OFFLINE) {
+          return { ...targetKind, path: '' }
+        }
+
+        return targetKind
+      })
+    )
+  }
+
+  const handleExternalOfflineKindPathOnFocus = () => {
+    setDisableButtons(state => ({
+      ...state,
+      isExternalOfflineTargetPathEditModeClosed: false
+    }))
+  }
+
+  const handleExternalOfflineKindPathOnBlur = ({ selectValue, inputValue }) => {
+    if (!isUrlInputValid(selectValue, inputValue, data[EXTERNAL_OFFLINE].kind)) {
+      setValidation(prevState => ({
+        ...prevState,
+        isExternalOfflineTargetPathValid: false
+      }))
+    } else {
+      if (!validation.isExternalOfflineTargetPathValid) {
+        setValidation(prevState => ({
+          ...prevState,
+          isExternalOfflineTargetPathValid: true
+        }))
+      }
+
       setNewFeatureSetTarget(
         featureStore.newFeatureSet.spec.targets.map(targetKind => {
           if (targetKind.name === EXTERNAL_OFFLINE) {
-            return { ...targetKind, path: event.target.value }
+            return { ...targetKind, path: `${selectValue}${inputValue}` }
           }
 
           return targetKind
         })
       )
     }
+
+    setData(state => ({
+      ...state,
+      [EXTERNAL_OFFLINE]: { ...state[EXTERNAL_OFFLINE], path: `${selectValue}${inputValue}` }
+    }))
+
+    setDisableButtons(state => ({
+      ...state,
+      isExternalOfflineTargetPathEditModeClosed: true
+    }))
   }
 
   const handleDiscardPathChange = kind => {
@@ -310,9 +354,8 @@ const FeatureSetsPanelTargetStore = ({
     setData(state => ({
       ...state,
       externalOffline: {
-        ...dataInitialState.externalOffline,
-        kind,
-        path: state.externalOffline.path
+        ...state.externalOffline,
+        kind
       }
     }))
     setNewFeatureSetTarget(
@@ -658,6 +701,7 @@ const FeatureSetsPanelTargetStore = ({
       handleAdvancedLinkClick={handleAdvancedLinkClick}
       handleDiscardPathChange={handleDiscardPathChange}
       handleExternalOfflineKindPathOnBlur={handleExternalOfflineKindPathOnBlur}
+      handleExternalOfflineKindPathOnFocus={handleExternalOfflineKindPathOnFocus}
       handleExternalOfflineKindTypeChange={handleExternalOfflineKindTypeChange}
       handleKeyBucketingNumberChange={handleKeyBucketingNumberChange}
       handleOfflineKindPathChange={handleOfflineKindPathChange}
@@ -667,6 +711,7 @@ const FeatureSetsPanelTargetStore = ({
       handlePartitionRadioButtonClick={handlePartitionRadioButtonClick}
       handleSelectTargetKind={handleSelectTargetKind}
       handleTimePartitioningGranularityChange={handleTimePartitioningGranularityChange}
+      handleUrlSelectOnChange={handleUrlSelectOnChange}
       partitionRadioButtonsState={partitionRadioButtonsState}
       frontendSpecIsNotEmpty={!isEmpty(frontendSpec.feature_store_data_prefixes)}
       selectedPartitionKind={selectedPartitionKind}
