+++ src/components/FeatureSetsPanel/FeatureSetsPanelTargetStore/featureSetsPanelTargetStore.scss
@@ -63,21 +63,10 @@
       margin-right: 5px;
     }
 
-    .input {
-      width: 100%;
-
-      &-wrapper {
-        &.url {
-          width: 300px;
-          margin-right: 5px;
-          margin-bottom: 5px;
-        }
-
-        &.partition-cols {
-          flex: unset;
-          width: 250px;
-        }
-      }
+    .url {
+      height: 52px;
+      margin-right: 5px;
+      margin-bottom: 5px;
     }
 
     .partition-fields {
