+++ src/elements/TableLinkCell/tableLinkCell.scss
@@ -55,9 +55,14 @@
 
     .status {
       display: flex;
+      flex: 0 1 20px;
       align-items: center;
       justify-content: center;
       margin-left: 5px;
+
+      & > * {
+        margin: 0;
+      }
     }
   }
 }
